Added one answer
-Unable to answer any of the other questions.  I completely mentally
froze, and my previous BDD repository was blank, and I barely was able
to absorb anything that we talked about with BDD.  I feel like I'm being
punished for trusting the instructors that we wouldn't be using this and
we were just "dipping our toe" into BDD; now I find that we're expected
to know exactly how to build BDD tests from scratch on our own.  I don't
even know where to start and I hate this.

Comfort: -1
Clarity: -1